Ladestationen overhaul

+++ src/main/java/controller/ParkhausServlet.java
@@ -63,11 +63,13 @@ public class ParkhausServlet extends HttpServlet {
     public static void doOnEveryRequest(HttpServletRequest req){
         TicketIF[] ladendeTickets = parkhaus.getLadendeTickets();
         TicketIF[] nichtLadendeTickets = parkhaus.getNichtLadendeTickets();
+
         TicketIF[] bezahlteTickets = parkhaus.getBezahlteTickets();
         TicketIF[] unbezahlteTickets = parkhaus.getUnbezahlteTickets();
+
         int freiePlaetze = parkhaus.getAnzahlFreiePlaetze();
-        float belegtePlaetze = bezahlteTickets.length + unbezahlteTickets.length;
-        int auslastung = (int) ((belegtePlaetze / (belegtePlaetze + freiePlaetze))*100);
+        int belegtePlaetze = bezahlteTickets.length + unbezahlteTickets.length;
+        int auslastung = (belegtePlaetze / (belegtePlaetze + freiePlaetze))*100;
 
         req.setAttribute("ladendeTickets", ladendeTickets);
         req.setAttribute("nichtLadendeTickets", nichtLadendeTickets);
